fix: checkout only if logged in
