+++ src/components/molecules/Login/googleLogout/GoogleLogout.js
@@ -1,19 +1,17 @@
 import React, { useContext } from "react";
-import { useGoogleLogin } from "react-google-login";
-import { textData } from "../../../../data/textData";
 import { ShopContext } from "../../../../context/ShopContext";
-// refresh token
-import { refreshTokenSetup } from "../../../../utils/refreshToken";
+import { useGoogleLogout } from "react-google-login";
+import { textData } from "../../../../data/textData";
 
 import styled from "styled-components";
 import { theme } from "../../../../data/theme";
 
-import ButtonX from "../../../subatoms/button/ButtonX";
+import ButtonX from "../../../atoms/button/ButtonX";
 
 const clientId =
   "589936013492-o5h98211ljn5r09rn4ih54203k4973fm.apps.googleusercontent.com";
 
-const StyledGoogleLogin = styled.div`
+const StyledGoogleLogout = styled.div`
   display: flex;
   gap: ${theme.spacer};
   justify-content: center;
@@ -28,50 +26,40 @@ const StyledGoogleLogin = styled.div`
   }
 `;
 
-const GoogleLogin = () => {
+const GoogleLogout = () => {
   const context = useContext(ShopContext);
 
-  const onSuccess = (res) => {
-    //alert(`Logged in successfully welcome ${res.profileObj.name} 😍.`);
-    refreshTokenSetup(res);
-    const name = res.profileObj.name;
+  const onLogoutSuccess = (res) => {
     context.setState({
       ...context.state,
-      googleUser: name,
-      signByGoogle: true,
-      isLoggedIn: true,
-      user: name,
+      googleUser: "",
+      signByGoogle: false,
+      isLoggedIn: !context.state.isLoggedIn,
       loginIconClicked: !context.state.loginIconClicked,
     });
   };
 
-  const onFailure = () => {
-    alert(`Failed to login. 😢`)
-  };
 
-  const { signIn } = useGoogleLogin({
-    onSuccess,
-    onFailure,
+  const { signOut } = useGoogleLogout({
     clientId,
-    isSignedIn: true,
-    accessType: "offline",
+    onLogoutSuccess,
   });
 
   return (
-    <StyledGoogleLogin>
+    <StyledGoogleLogout>
       <img
         src="https://raw.githubusercontent.com/Sivanesh-S/react-google-authentication/9835990bfe7f45a1a14e5854cd57ab715d776b0e/public/icons/google.svg"
         alt="google login"
-        onClick={signIn}
+        onClick={signOut}
       />
       <ButtonX
-        action={signIn}
+        action={signOut}
         size="M"
         color="light"
-        text={textData.login.text2}
+        text={textData.logout.button}
       />
-    </StyledGoogleLogin>
+    </StyledGoogleLogout>
   );
 };
 
-export default GoogleLogin;
+export default GoogleLogout;
