+++ src/components/molecules/Login/login-fb/FacebookLogin.js
@@ -1,12 +1,14 @@
-import React, { useContext } from "react";
+import React, { useState, useEffect, useContext } from "react";
+import { accountService } from "../../../../_services/account.service";
 import { ShopContext } from "../../../../context/ShopContext";
 import { textData } from "../../../../data/textData";
+
 import styled from "styled-components";
 import { theme } from "../../../../data/theme";
-import ButtonX from "../../../subatoms/button/ButtonX";
-import { accountService } from "../../../../_services/account.service";
 
-const StyledFacebookLogout = styled.div`
+import ButtonX from "../../../atoms/button/ButtonX";
+
+const StyledFacebookLogin = styled.div`
   display: flex;
   gap: ${theme.spacer};
   justify-content: center;
@@ -16,40 +18,44 @@ const StyledFacebookLogout = styled.div`
   img {
     width: ${theme.sizes.buttons.M};
     height: ${theme.sizes.buttons.M};
-    background-color: ${theme.colors.light};
-    border-radius: 50%;
+    border-radius: 10%;
   }
 `;
 
-const FacebookLogout = () => {
+const FacebookLogin = () => {
   const context = useContext(ShopContext);
 
-  const logout = () => {
-    accountService.logout();
+  const [account, setAccount] = useState(null);
+
+  useEffect(() => {
+    accountService.account.subscribe((x) => setAccount(x));
+  }, []);
+
+  if (account) {
     context.setState({
       ...context.state,
+      signByFB: true,
       isLoggedIn: !context.state.isLoggedIn,
-      user: "",
+      user: account.name,
       loginIconClicked: !context.state.loginIconClicked,
-      signByFB: false,
     });
-  };
+  }
 
   return (
-    <StyledFacebookLogout>
+    <StyledFacebookLogin>
       <img
         src="https://upload.wikimedia.org/wikipedia/commons/thumb/f/fb/Facebook_icon_2013.svg/1200px-Facebook_icon_2013.svg.png"
         alt="facebook login"
-        className="icon"
+        onClick={accountService.login}
       />
       <ButtonX
-        action={logout}
+        action={accountService.login}
         size="M"
         color="light"
-        text={textData.logout.button}
+        text={textData.login.text3}
       />
-    </StyledFacebookLogout>
+    </StyledFacebookLogin>
   );
 };
 
-export default FacebookLogout;
+export default FacebookLogin;
