+++ src/components/molecules/Login/loginModal/LoginModal.js
@@ -5,10 +5,10 @@ import { Link } from "react-router-dom";
 import { ShopContext } from "../../../../context/ShopContext";
 import { textData } from "../../../../data/textData";
 import { LogoImg } from "../../mainNavbar/navbar/Navbar";
-import Text from "../../../subatoms/text/Text";
+import Text from "../../../atoms/text/Text";
 import logo from "./../../../../data/images/logonorris.png";
-import ButtonX from "../../../subatoms/button/ButtonX";
-import LoginInput from "../../../subatoms/input/LoginInput";
+import ButtonX from "../../../atoms/button/ButtonX";
+import LoginInput from "../../../atoms/input/LoginInput";
 import GoogleLogin from "../googleLogin/GoogleLogin";
 import FacebookLogin from "../login-fb/FacebookLogin";
