+++ src/components/pages/shop/Shop.js
@@ -7,10 +7,10 @@ import { textData } from "../../../data/textData";
 import ShopBanner from "./shop-banner/ShopBanner";
 import SingleShop from "./remote-shop/SingleShop";
 import ShopIconWrap from "./shop-icon-wrap/ShopIconWrap";
-import Spinner from "../../subatoms/spinner";
+import Spinner from "../../atoms/spinner";
 import SingleShopCard from "./remote-shop/SingleShopCard";
-import MainNavbar from "../../atoms/mainNavbar/MainNavbar";
-import Footer from "./../../atoms/footer/Footer";
+import MainNavbar from "../../molecules/mainNavbar/MainNavbar";
+import Footer from "./../../molecules/footer/Footer";
 
 import jewerly from "./../../../data/images/jewelry.png";
 import jacket from "./../../../data/images/jacket.png";
