+++ src/components/pages/shop/remote-shop/SingleShopCard.js
@@ -1,5 +1,5 @@
-import Text from "./../../../subatoms/text/Text";
-import Button from "../../../subatoms/button/Button";
+import Text from "./../../../atoms/text/Text";
+import Button from "../../../atoms/button/Button";
 import { textData } from "../../../../data/textData";
 import { theme } from "../../../../data/theme";
 import styled from "styled-components";
