+++ src/components/pages/shop/stripe-button/StripeCheckoutButton.js
@@ -3,7 +3,7 @@ import StripeCheckout from "react-stripe-checkout";
 import { MyContext } from "../../../../context/APIContext";
 import { textData } from "../../../../data/textData";
 
-import Button from "../../../subatoms/button/Button";
+import Button from "../../../atoms/button/Button";
 
 const StripeCheckoutButton = ({ price }) => {
   const context = useContext(MyContext);
@@ -13,7 +13,6 @@ const StripeCheckoutButton = ({ price }) => {
     "pk_test_51HyOJsGQ9zu5MlObNnBhMMgC5EQpWd8sCE0mZGmeoOL4J9teVtSFrLWYuRJOCA2C75iYqQJaOuv4Fjp2ymiXGpWZ00oZn7L6BB";
 
   const onToken = (token) => {
-    console.log(token);
     alert("Payment Succesful!");
     context.setCart("");
   };
